Refactor main python
